gpu/execute: fix vs_identity reference type to uint64_t in LayeredVolumeSynthesisOutput

+++ prosper/src/gpu/execute/layered_volume_target.hpp
@@ -180,7 +180,7 @@ struct LayeredVolumeSynthesisOutput {
     std::vector<uint32_t>& vs;
     std::vector<uint32_t>& gs;
     std::shared_ptr<const std::vector<uint32_t>>& vs_shared;
-    uint32_t& vs_identity;
+    uint64_t& vs_identity;
     std::shared_ptr<ShaderResourceTable>& vrt;
 };
